Reorganize the signing-related code
Here we reorganize the code around signing to improve readability and overall
code structure. Specifically, we now have the following files holding the
code related to signing:
- signing_controller.go: Holds everything related to the top-level signing
  group controller
- signing_loop.go: Holds everything related to the signing retry loop
- signing_syncer.go: Holds everything related to synchronization upon
  a successful signing
